requirements.txt
alembic==1.13.2
annotated-types==0.7.0
anyio==3.6.2
astroid==3.2.4
async-timeout==4.0.3
asyncpg==0.29.0
cffi==1.15.1
click==8.1.3
cryptography==40.0.1
dill==0.3.6
ecdsa==0.18.0
fastapi==0.112.1
greenlet==2.0.2
h11==0.14.0
httptools==0.5.0
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
Mako==1.2.4
MarkupSafe==2.1.2
mccabe==0.7.0
passlib==1.7.4
pep8==1.7.1
platformdirs==3.2.0
psycopg2-binary==2.9.6
pyasn1==0.4.8
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
pylint==3.2.6
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.32
starlette==0.38.2
tomlkit==0.11.7
typing_extensions==4.12.2
uvicorn==0.21.1
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.1
wrapt==1.15.0
